+++ src/main/java/TopChefRPG/model/User.java
@@ -8,7 +8,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Long id;
+    private int id;
 
     @Column(name="name")
     private String name;
@@ -38,7 +38,7 @@ public class User {
         return name;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
